% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_convs.R
\name{boof_buac_to_kgha_corn}
\alias{boof_buac_to_kgha_corn}
\title{boof_buac_to_kgha_corn}
\usage{
boof_buac_to_kgha_corn(dat_buac)
}
\arguments{
\item{dat_buac}{A vector of corn yields in bushels per acre}
}
\value{
A vector of corn yields in dry kilograms per hectare; adjusted for the 15\% moisture the bu/ac are reported at
}
\description{
boof_buac_to_kgha_corn
}
