% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day-of-year_convs.R
\name{boof_date_to_doy}
\alias{boof_date_to_doy}
\title{boof_date_to_doy}
\usage{
boof_date_to_doy(mydate = "2001-01-01")
}
\arguments{
\item{mydate}{A date in yyyy-mm-dd format}
}
\value{
A numeric value representing the day of year, doy
}
\description{
boof_date_to_doy
}
