% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day-of-year_convs.R
\name{boof_doy_to_date}
\alias{boof_doy_to_date}
\title{boof_doy_to_date}
\usage{
boof_doy_to_date(mydoy = 1, myyear = 2000)
}
\arguments{
\item{mydoy}{A day of year}

\item{myyear}{A year if known, defaults to 2000}
}
\value{
A date in the format yyyy-mm-dd
}
\description{
boof_doy_to_date
}
