% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_convs.R
\name{boof_kgha_to_buac_corn}
\alias{boof_kgha_to_buac_corn}
\title{boof_kgha_to_buac_corn}
\usage{
boof_kgha_to_buac_corn(dat_kgha)
}
\arguments{
\item{dat_kgha}{A vector of corn yields in DRY kgs per hectare}
}
\value{
A vector of corn yields in bushels per acre; adjusted for the 15.5\% moisture the bu/ac are reported at
}
\description{
boof_kgha_to_buac_corn
}
