% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs_wrap.R
\name{boof_labs_wrap}
\alias{boof_labs_wrap}
\title{Wrap ggplot labels}
\usage{
boof_labs_wrap(..., width = 80)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:labs]{ggplot2::labs()}}}

\item{width}{The width of the characters to wrap to.}
}
\description{
\code{boof_labs_wrap()} wraps \code{\link[stringr:str_wrap]{stringr::str_wrap()}} around any argument passed to
\code{\link[ggplot2:labs]{ggplot2::labs()}}, thus wrapping it. Taken from \href{https://malco.io/}{Malcolm Barrett}
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, hp)) +
boof_labs_wrap(title =
  "Here is my really long title. You see, I have a lot to say, you see.",
  width = 30)
}
