% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_convs.R
\name{boof_lbac_to_kgha}
\alias{boof_lbac_to_kgha}
\title{boof_lbac_to_kgha}
\usage{
boof_lbac_to_kgha(dat_lbac)
}
\arguments{
\item{dat_lbac}{A vector of values in pounds per acre}
}
\value{
A vector of values in kilograms per hectare; multiplied by 1.123
}
\description{
boof_lbac_to_kgha
}
